Add files via upload
This is the group report that should be later emailed to the teacher as the attachment and added to our project's folder.  I started it, but did not fill it out completely because I don't have the full understanding of everything that we did.  Please make sure you add your contribution/focus at the bottom of the page as well.